requirements.txt
dnspython==2.7.0
Faker==37.6.0
numpy==2.3.2
pandas==2.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
